requirements.txt
annotated-types==0.6.0
anyio==4.3.0
arrow==1.3.0
binaryornot==0.4.4
blessed==1.20.0
CacheControl==0.14.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cookiecutter==2.6.0
cryptography==42.0.5
customtkinter==5.2.2
darkdetect==0.8.0
decorator==5.1.1
editor==1.6.6
fastapi==0.110.0
firebase-admin==6.5.0
future==1.0.0
geocoder==1.38.1
geographiclib==2.0
geopy==2.4.1
gmplot==1.4.1
google-api-core==2.18.0
google-api-python-client==2.124.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.15.0
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpcio==1.62.1
grpcio-status==1.62.1
h11==0.14.0
httpcore==1.0.4
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
idna==3.6
inquirer==3.2.4
Jinja2==3.1.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
msgpack==1.0.8
oauthlib==3.2.2
packaging==23.2
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycountry==23.12.11
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.17.2
PyJWT==2.8.0
pyparsing==3.1.2
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
python-telegram-bot==21.0.1
PyYAML==6.0.1
qrcode==7.4.2
ratelim==0.1.6
readchar==4.0.6
repath==0.9.0
requests==2.31.0
rich==13.7.1
rsa==4.9
runs==1.2.2
setuptools==69.2.0
simple_chalk==0.1.0
six==1.16.0
sniffio==1.3.1
starlette==0.36.3
terminaltexteffects==0.11.0
text-unidecode==1.3
tk==0.1.0
tqdm==4.66.2
types-python-dateutil==2.9.0.20240316
typing_extensions==4.10.0
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.28.1
uvloop==0.19.0
watchdog==4.0.0
watchfiles==0.21.0
wcwidth==0.2.13
websockets==12.0
xmod==1.8.1
